move crc32 in internal module
