+++ wallet-crypto/src/lib.rs
@@ -1,3 +1,4 @@
+mod crc32;
 pub mod cbor;
 pub mod hdwallet;
 pub mod paperwallet;
